+++ data/source/service/client/client.go
@@ -1,4 +1,4 @@
-package service
+package client
 
 import (
 	"context"
@@ -11,22 +11,22 @@ import (
 	"github.com/tidepool-org/platform/permission"
 )
 
-type ClientProvider interface {
+type Provider interface {
 	AuthClient() auth.Client
 	DataSourceStructuredStore() dataSourceStoreStructured.Store
 }
 
 type Client struct {
-	ClientProvider
+	Provider
 }
 
-func NewClient(clientProvider ClientProvider) (*Client, error) {
-	if clientProvider == nil {
-		return nil, errors.New("client provider is missing")
+func New(provider Provider) (*Client, error) {
+	if provider == nil {
+		return nil, errors.New("provider is missing")
 	}
 
 	return &Client{
-		ClientProvider: clientProvider,
+		Provider: provider,
 	}, nil
 }